allocate_displaced_instruction: guide template argument
In aql_queue_t::allocate_displaced_instruction, MSVC complains about an
initialization of a uint16_t using a size_t.  Explicitly cast the
argument to uint16_t before calling emplace so make things explicit.

The code has assertions in place to ensure that overflow cannot happen.

+++ src/queue.cpp
@@ -472,7 +472,7 @@ aql_queue_t::allocate_displaced_instruction (const instruction_t &instruction)
       = (ptr - *m_debugger_memory_base) / debugger_memory_chunk_size;
 
     dbgapi_assert (index < m_debugger_memory_chunk_count);
-    m_debugger_memory_free_chunks.emplace_back (index);
+    m_debugger_memory_free_chunks.emplace_back (static_cast<uint16_t> (index));
   };
 
   dbgapi_assert (instruction.size () + assert_instruction.size ()
